pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-recall"
version = "0.5.0"
description = "Four-layer cognitive memory for AI agents — persistent recall, semantic search, knowledge graph, and learned evolution. No API key required."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jonathan Domian", email = "jonathan.domian@gmail.com"},
]
keywords = [
    "ai",
    "agents",
    "memory",
    "cognitive",
    "rag",
    "knowledge-graph",
    "semantic-search",
    "vector-database",
    "chromadb",
    "llm",
    "claude",
    "mcp",
    "embeddings",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "chromadb>=0.5.0,<0.7",
    "pyyaml>=6.0,<7",
    "sentence-transformers>=2.2,<4",
]

[project.optional-dependencies]
dev = [
    "build",
    "twine",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]
spellcheck = [
    "pyspellchecker>=0.7.0",
]

[project.scripts]
cortex = "cortex.cli:main"

[project.urls]
Homepage = "https://github.com/jdomian/cortex-ai"
Repository = "https://github.com/jdomian/cortex-ai"
"Bug Tracker" = "https://github.com/jdomian/cortex-ai/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cortex*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
"cortex" = ["instructions/*.md", "py.typed"]
